Renamed and reorganized classes
