+++ src/main/java/code/elix_x/excore/utils/client/render/pipeline/model/ModelQuadsPipeline.java
@@ -1,4 +1,4 @@
-package code.elix_x.excore.utils.client.render.pipeline;
+package code.elix_x.excore.utils.client.render.pipeline.model;
 
 import code.elix_x.excomms.pipeline.PipelineElement;
 import code.elix_x.excomms.pipeline.list.ListPipelineElement;
@@ -6,11 +6,11 @@ import code.elix_x.excore.utils.client.render.model.UnpackedBakedQuad;
 import code.elix_x.excore.utils.client.render.model.UnpackedSimpleBakedModel;
 import net.minecraft.util.EnumFacing;
 
-public class BakedModelQuadsPipelineWrapper implements PipelineElement<UnpackedSimpleBakedModel, UnpackedSimpleBakedModel> {
+public class ModelQuadsPipeline implements PipelineElement<UnpackedSimpleBakedModel, UnpackedSimpleBakedModel> {
 
 	private final ListPipelineElement<UnpackedBakedQuad, UnpackedBakedQuad> pipeline;
 
-	public BakedModelQuadsPipelineWrapper(ListPipelineElement<UnpackedBakedQuad, UnpackedBakedQuad> pipeline){
+	public ModelQuadsPipeline(ListPipelineElement<UnpackedBakedQuad, UnpackedBakedQuad> pipeline){
 		this.pipeline = pipeline;
 	}
