+++ src/main/java/code/elix_x/excore/utils/client/render/pipeline/model/Packer.java
@@ -1,11 +1,11 @@
-package code.elix_x.excore.utils.client.render.pipeline;
+package code.elix_x.excore.utils.client.render.pipeline.model;
 
 import code.elix_x.excomms.pipeline.PipelineElement;
 import code.elix_x.excore.utils.client.render.model.UnpackedSimpleBakedModel;
 import net.minecraft.client.renderer.block.model.SimpleBakedModel;
 import net.minecraft.client.renderer.vertex.VertexFormat;
 
-public class PackerPipelineElement {
+public class Packer {
 
 	public static PipelineElement<SimpleBakedModel, UnpackedSimpleBakedModel> unpack(){
 		return PipelineElement.wrapper(UnpackedSimpleBakedModel::unpack);
