+++ src/main/java/code/elix_x/excore/utils/client/render/pipeline/quad/QuadVerticesPipeline.java
@@ -1,14 +1,14 @@
-package code.elix_x.excore.utils.client.render.pipeline;
+package code.elix_x.excore.utils.client.render.pipeline.quad;
 
 import code.elix_x.excomms.pipeline.PipelineElement;
 import code.elix_x.excore.utils.client.render.model.UnpackedBakedQuad;
 import code.elix_x.excore.utils.client.render.model.vertex.DefaultUnpackedVertices;
 
-public class BakedQuadVerticesPipelineWrapper implements PipelineElement<UnpackedBakedQuad, UnpackedBakedQuad> {
+public class QuadVerticesPipeline implements PipelineElement<UnpackedBakedQuad, UnpackedBakedQuad> {
 
 	private final PipelineElement<DefaultUnpackedVertices, DefaultUnpackedVertices> pipeline;
 
-	public BakedQuadVerticesPipelineWrapper(PipelineElement<DefaultUnpackedVertices, DefaultUnpackedVertices> pipeline){
+	public QuadVerticesPipeline(PipelineElement<DefaultUnpackedVertices, DefaultUnpackedVertices> pipeline){
 		this.pipeline = pipeline;
 	}
